refactored dominance result -> model
