+++ src/main/java/gr/uoi/cs/pythia/model/PatternsProfile.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
+import gr.uoi.cs.pythia.model.dominance.DominanceResult;
 import gr.uoi.cs.pythia.model.outlier.OutlierResult;
-import gr.uoi.cs.pythia.patterns.dominance.DominanceResult;
 
 public class PatternsProfile {
