+++ src/main/java/gr/uoi/cs/pythia/patterns/PatternManager.java
@@ -10,11 +10,11 @@ import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 
 import gr.uoi.cs.pythia.model.DatasetProfile;
+import gr.uoi.cs.pythia.model.dominance.DominanceResult;
 import gr.uoi.cs.pythia.model.outlier.OutlierResult;
 import gr.uoi.cs.pythia.model.outlier.OutlierType;
 import gr.uoi.cs.pythia.patterns.dominance.DominanceColumnSelector;
 import gr.uoi.cs.pythia.patterns.dominance.DominanceParameters;
-import gr.uoi.cs.pythia.patterns.dominance.DominanceResult;
 import gr.uoi.cs.pythia.patterns.dominance.HighDominanceAlgo;
 import gr.uoi.cs.pythia.patterns.dominance.LowDominanceAlgo;
 import gr.uoi.cs.pythia.patterns.outlier.IOutlierAlgo;
